requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
MarkupSafe==3.0.2
pigpio==1.78
prometheus_client==0.21.1
PyYAML==6.0.2
rpi-lgpio==0.6
RPi.GPIO==0.7.1
smbus==1.1.post2
Werkzeug==3.1.3
